Move Translate specific components
