+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/translate/components/DownloadCSV.tsx
@@ -1,11 +1,11 @@
 "use client";
 import React from "react";
 import { useTranslation } from "@i18n/client";
-import { useTemplateStore } from "../../store/useTemplateStore";
+import { useTemplateStore } from "@clientComponents/form-builder/store/useTemplateStore";
 import { Button } from "@clientComponents/globals";
 import { FormElement } from "@lib/types";
 import { getDate, slugify } from "@lib/client/clientHelpers";
-import { alphabet } from "../../util";
+import { alphabet } from "@clientComponents/form-builder/util";
 
 const formatText = (str: string) => `"${str}"`;
