+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/translate/components/FieldsetLegend.tsx
@@ -3,7 +3,7 @@ import React from "react";
 
 export const FieldsetLegend = ({ children }: { children: React.ReactNode }) => {
   return (
-    <legend className="text-sm p-2 border border-b-0 border-gray-300 bg-gray-100 w-full">
+    <legend className="w-full border border-b-0 border-gray-300 bg-gray-100 p-2 text-sm">
       {children}
     </legend>
   );
