+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/translate/components/index.ts
