+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/translate/page.tsx
@@ -1,6 +1,6 @@
 import { serverTranslation } from "@i18n";
 import { EditNavigation } from "@clientComponents/form-builder/app";
-import { Translate } from "@clientComponents/form-builder/app/translate";
+import { Translate } from "app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/translate/components";
 import { Metadata } from "next";
 
 export async function generateMetadata({
