--- components/clientComponents/form-builder/app/translate/SectionTitle.tsx
@@ -1,11 +1,0 @@
-"use client";
-import React from "react";
-
-export const SectionTitle = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <div className="mt-2 mb-4 flex">
-      <h2 className="text-[24px] m-0 p-0 pr-4">{children}</h2>
-      <hr className="border mt-5 border-black border-dotted flex grow" />
-    </div>
-  );
-};
